pyproject.toml
[project]
name = "python3-lgpd-crawler"
version = "0.1.0"
description = "App de sondagem, apreciação e mapeamento dos dados pessoais e/ou sensíveis para busca de violações de compliance de acordo com as recomendações do LGPD (adaptável mediante dicionário de regras)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc>=0.0.4",
    "annotated-types>=0.7.0",
    "anyio>=4.12.1",
    "asyncio>=4.0.0",
    "blinker==1.9.0",
    "certifi==2025.11.12",
    "cffi==2.0.0 ; platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "contourpy>=1.3.3",
    "cryptography==46.0.5",
    "cycler>=0.12.1",
    "defusedxml>=0.7.1",
    "et-xmlfile>=2.0.0",
    "extract-msg>=0.55.0",
    "fastapi>=0.129.2",
    "flask==3.1.3",
    "flask-sqlalchemy==3.1.1",
    "fonttools>=4.61.1",
    "greenlet==3.3.0 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "idna==3.11",
    "iniconfig>=2.3.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "joblib>=1.5.3",
    "kiwisolver>=1.4.9",
    "mariadb==1.1.14",
    "markupsafe==3.0.3",
    "matplotlib>=3.10.8",
    "mysql==0.0.3",
    "mysqlclient==2.2.7",
    "numpy>=2.4.2",
    "odfpy>=1.4.1",
    "openpyxl>=3.1.5",
    "oracledb==3.4.1",
    "packaging==25.0",
    "pandas>=3.0.1",
    "pillow==12.1.1",
    "pluggy>=1.6.0",
    "psycopg2-binary==2.9.11",
    "pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'",
    "pydantic>=2.12.5",
    "pydantic-core>=2.41.5",
    "pygments>=2.19.2",
    "pymysql>=1.1.2",
    "pyodbc>=5.3.0",
    "pyparsing>=3.3.2",
    "pypdf>=6.7.2",
    "pypdf2>=3.0.1",
    "pytest>=9.0.2",
    "python-dateutil>=2.9.0.post0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "pyyaml>=6.0.3",
    "regex>=2026.2.19",
    "reportlab==4.4.7",
    "requests==2.32.5",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "seaborn>=0.13.2",
    "setuptools==80.9.0",
    "six>=1.17.0",
    "sqlalchemy==2.0.45",
    "starlette>=0.52.1",
    "threadpoolctl>=3.6.0",
    "typing-extensions==4.15.0",
    "typing-inspection>=0.4.2",
    "urllib3==2.6.3",
    "uvicorn>=0.41.0",
    "werkzeug==3.1.6",
]
